requirements.txt
pygame==2.6.1
Pillow==9.3.0
